# Avoid lock contention for error backoff