#### What type of PR is this?

/kind cleanup

#### What this PR does / why we need it:

Avoids lock contention for error backoff.

I realize this is a minor fix because this operation sleeps, but it shows up on lock contention detection tools and I'd like to clear it.

BEFORE:

BenchmarkRudimentaryErrorBackoff_OnError
BenchmarkRudimentaryErrorBackoff_OnError/Goroutines-1
BenchmarkRudimentaryErrorBackoff_OnError/Goroutines-1-14         	 2219292	       490.1 ns/op
BenchmarkRudimentaryErrorBackoff_OnError/Goroutines-10
BenchmarkRudimentaryErrorBackoff_OnError/Goroutines-10-14        	  102898	     12251 ns/op
BenchmarkRudimentaryErrorBackoff_OnError/Goroutines-100
BenchmarkRudimentaryErrorBackoff_OnError/Goroutines-100-14       	    2866	    492000 ns/op
BenchmarkRudimentaryErrorBackoff_OnError/Goroutines-1000
BenchmarkRudimentaryErrorBackoff_OnError/Goroutines-1000-14      	     684	   1772103 ns/op
PASS

AFTER:

BenchmarkRudimentaryErrorBackoff_OnError
BenchmarkRudimentaryErrorBackoff_OnError/Goroutines-1
BenchmarkRudimentaryErrorBackoff_OnError/Goroutines-1-14         	 2317406	       470.3 ns/op
BenchmarkRudimentaryErrorBackoff_OnError/Goroutines-10
BenchmarkRudimentaryErrorBackoff_OnError/Goroutines-10-14        	  380134	      3210 ns/op
BenchmarkRudimentaryErrorBackoff_OnError/Goroutines-100
BenchmarkRudimentaryErrorBackoff_OnError/Goroutines-100-14       	   31053	     37920 ns/op
BenchmarkRudimentaryErrorBackoff_OnError/Goroutines-1000
BenchmarkRudimentaryErrorBackoff_OnError/Goroutines-1000-14      	    3194	    370947 ns/op
PASS

#### Does this PR introduce a user-facing change?

```release-note
NONE
```